Refactored with hylo-idl crate, isolating IDL from core.
