+++ Cargo.toml
@@ -1,7 +1,7 @@
 [workspace]
 members = [
   "hylo-clients",
-  "hylo-core",
+  "hylo-core", "hylo-idl", "hylo-jupiter",
 ]
 resolver = "2"
 
@@ -26,3 +26,5 @@ futures = "0.3.31"
 itertools = "0.14.0"
 jupiter-amm-interface = "0.6.0"
 proptest = "1.5.0"
+rust_decimal = "1.37.2"
+bincode = "1.3.3"
