+++ hylo-clients/Cargo.toml
@@ -5,14 +5,14 @@ edition.workspace = true
 description = "Hylo RPC clients"
 license.workspace = true
 homepage.workspace = true
-include = ["idls/", "src/"]
 
 [dependencies]
 anyhow.workspace = true
 anchor-client.workspace = true
 anchor-spl.workspace = true
 anchor-lang.workspace = true
-hylo-core = { path = "../hylo-core" }
+hylo-core = { path = "../hylo-core", features = ["offchain"] }
+hylo-idl = { path = "../hylo-idl" }
 hylo-fix.workspace = true
 pyth-solana-receiver-sdk.workspace = true
 base64.workspace = true
@@ -21,6 +21,3 @@ tokio.workspace = true
 async-trait.workspace = true
 futures.workspace = true
 itertools.workspace = true
-jupiter-amm-interface.workspace = true
-rust_decimal = "1.37.2"
-bincode = "1.3.3"
