+++ hylo-clients/src/lib.rs
@@ -1,23 +1,3 @@
-#![allow(clippy::pub_underscore_fields)]
-
-extern crate anchor_lang;
-
-anchor_lang::declare_program!(hylo_exchange);
-anchor_lang::declare_program!(hylo_stability_pool);
-
-mod exchange_client;
-pub mod idl_type_bridge;
-pub mod jupiter;
-pub mod pda;
-mod stability_pool_client;
+pub mod exchange_client;
+pub mod stability_pool_client;
 pub mod util;
-
-pub mod exchange {
-  pub use super::exchange_client::*;
-  pub use super::hylo_exchange::*;
-}
-
-pub mod stability_pool {
-  pub use super::hylo_stability_pool::*;
-  pub use super::stability_pool_client::*;
-}
