+++ hylo-core/src/lib.rs
@@ -3,6 +3,8 @@ pub mod error;
 pub mod exchange_context;
 pub mod exchange_math;
 pub mod fee_controller;
+#[cfg(feature = "offchain")]
+pub mod idl_type_bridge;
 pub mod lst_sol_price;
 pub mod pyth;
 pub mod slippage_config;
@@ -11,3 +13,6 @@ pub mod stability_mode;
 pub mod stability_pool_math;
 pub mod total_sol_cache;
 mod util;
+
+#[cfg(feature = "offchain")]
+pub use hylo_idl as idl;
