+++ hylo-idl/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "hylo-idl"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+homepage.workspace = true
+include = ["idls/", "src/"]
+
+[dependencies]
+anchor-lang.workspace = true
+anchor-spl.workspace = true
