+++ hylo-idl/idls/hylo_stability_pool.json
